+++ vales/apps/maestros/models/socio_models.py
@@ -15,12 +15,12 @@ from entorno.constantes_base import (SOLICITUD_SOCIO,
 class Socio(ModeloBaseGenerico):
 	id_socio = models.AutoField(primary_key=True)
 	estatus_socio = models.BooleanField("Estatus*", default=False, 
-										  choices=ESTATUS_GEN)
+									choices=ESTATUS_GEN)
 	codigo_socio = models.IntegerField("Código", null=True, blank=True)
 	nombre_socio = models.CharField("Nombre Socio*", max_length=50)
 	domicilio_socio = models.CharField("Domicilio Socio*", max_length=50)
 	codigo_postal = models.CharField("Código Postal*", max_length=5,
-                                  null=True, blank=True)
+                                    null=True, blank=True)
 	id_provincia = models.ForeignKey(Provincia, on_delete=models.PROTECT, 
 									verbose_name="Provincia*",
           							null=True, blank=True)
@@ -33,37 +33,37 @@ class Socio(ModeloBaseGenerico):
 	id_tipo_iva = models.ForeignKey(TipoIva, on_delete=models.PROTECT,
 									verbose_name="Tipo de IVA*")
 	id_tipo_documento_identidad = models.ForeignKey(TipoDocumentoIdentidad, 
-										on_delete=models.PROTECT, 
-										verbose_name="Tipo Doc. Identidad*")
+									on_delete=models.PROTECT, 
+									verbose_name="Tipo Doc. Identidad*")
 	numero_documento = models.IntegerField("Número doc.*", null=True, blank=True)
-	cuit = models.IntegerField("CUIT/CUIL*", unique=True, null=True, blank=True)
+	cuit = models.BigIntegerField("CUIT/CUIL*", unique=True, null=True, blank=True)
 	telefono_socio = models.CharField("Teléfono*", max_length=15)
 	telefono2_socio = models.CharField("Teléfono Alternativo", max_length=15,
-									 null=True, blank=True)
+									null=True, blank=True)
 	movil_socio = models.CharField("Móvil", max_length=15, null=True, blank=True)
 	email_socio = models.EmailField("Email*", max_length=50)
 	fecha_nacimiento = models.DateField("Fecha Nacimiento", 
-									 null=True, blank=True)
+									null=True, blank=True)
 	fecha_alta = models.DateField("Fecha Alta", default=date.today,
-                               null=True, blank=True)
+                               		null=True, blank=True)
 	sexo = models.CharField("Sexo*", max_length=1, 
-							default="M", choices=SEXO)
+									default="M", choices=SEXO)
 	id_sucursal = models.ForeignKey(Sucursal, 
 									on_delete=models.CASCADE,
 									null=True, blank=True,
 									verbose_name="Sucursal*")
 	limite_credito = models.DecimalField("Límite de Crédito", 
-										 max_digits=15, decimal_places=2,
-										 default=0.00)
+									max_digits=15, decimal_places=2,
+									default=0.00)
 	disponible_credito = models.DecimalField("Crédito Disponible", 
-											max_digits=15, decimal_places=2,
-											default=0.00)
+									max_digits=15, decimal_places=2,
+									default=0.00)
 	black_list = models.BooleanField("Black List", default=False, 
-										  choices=SI_NO)
+									 choices=SI_NO)
 	black_list_motivo = models.CharField("Motivo Black List", max_length=50, 
-										   null=True, blank=True)
+									  null=True, blank=True)
 	black_list_usuario = models.CharField("Usuario Black List", 
-										  max_length=20, null=True, blank=True)
+									 max_length=20, null=True, blank=True)
 	fecha_baja = models.DateField("Fecha de Baja", null=True, blank=True)
 	class Meta:
 		db_table = 'socio'
@@ -125,7 +125,7 @@ class SolicitudAdhesion(ModeloBaseGenerico):
 									choices=ESTATUS_GEN)
 	id_socio = models.ForeignKey(Socio, on_delete=models.CASCADE,
 									verbose_name="Socio*")
-	cuit_solicitud_adhesion = models.IntegerField("CUIT/CUIL*", 
+	cuit_solicitud_adhesion = models.BigIntegerField("CUIT/CUIL*", 
 									null=True, blank=True)
 	limite_credito_solicitud_adhesion = models.DecimalField("Límite de Crédito Solicitado", 
 									max_digits=15, decimal_places=2,
